requirements.txt
pandas
numpy
snowflake-snowpark-python
snowflake-connector-python
matplotlib
plotly
scikit-learn
python-dotenv
